server: inbox / login の HTML 生成を inbox-view.js へ切り出す (#109 R-2)
挙動を変えないリファクタリング(Issue #109 スライス R-2)。HTML 出力は
バイト単位で不変(inbox の data 属性・削除ボタン等を pin する既存テストは
無変更で pass)。

- renderInbox / renderImportPanel / renderFilterPanel / renderGitHubCell /
  renderScreenshotPreview / formatSlackStatus / feedbackStatusOf /
  renderLoginPage を server/inbox-view.js(CJS)へ verbatim move
- inbox-view.js は shared/format.js と store.js の FEEDBACK_STATUSES を
  直接 require し、receiver の設定・状態に依存するもの(screenshotUrlFor /
  formatScreenshotStatus / safeLinkUrl / GITHUB_CONFIGURED 等)は
  createInboxView(deps) factory で注入
- responder(handleGetInbox / respondLoginPage 等)と Slack / GitHub 整形と
  共有するヘルパは receive.js に残置

実装: Codex(Claude 作成のブリーフに基づく)。検証: npm run check 全 113 テスト pass。
